[manager/orchestrator/taskreaper] Move task reaper tests to orchestrator/taskreaper
(cherry picked from commit 8cfb337920a6658b302643f27074ca3d669176ec)
